Incremental_Pipeline/requirements.txt
google-cloud-bigquery==3.25.0
google-auth==2.35.0
google-cloud-storage==2.18.2
requests==2.32.3
tenacity==9.0.0
urllib3==2.2.3
python-dotenv==1.0.1
pytest
